- Implementacion del Ejercicio #5 del Capitulo 5 (Ventana emergente, programacion de funcionalidad de cada boton y textbox)
- Implementacion de Clase (ConvertirALetras) para el Ejercicio #5 del Capitulo 5

- Correccion de errores en titulos de ventanas emergentes

- Cambio de formato en el texto de los resultados de los textbox

+++ EjerciciosCap4y5/Ejercicios_Capitulo_4/Ejercicio2_Cap4.xaml.cs
@@ -24,8 +24,8 @@ namespace EjerciciosCap4y5
 
         private void ButtonClick_Ejercicio2_Cap4(object sender, RoutedEventArgs e)
         {
-            int numero;
-            int potencia;
+            double numero;
+            double potencia;
             numero = int.Parse(Numero2.Text);
             potencia = int.Parse(Potencia.Text);
             double elevado = Math.Pow(numero, potencia);
